requirements.txt
 tkinter 
  
 threading 
 cv2
 imutils
 scipy.spatial 
 
 
 dlib
 pygame 